Implemented template based export.
Executable includes the default template which produces the usual data, but user can inject their own. The order in which templates are takes is:

1. Template file is found at the location of project XML (name must be "<project filename>.template"
2. Template file is found at the location of the NextGraphics.exe (name must be "NextGraphics.template"
3. Built in template is used

Note: care was taken while creating default template to make it as close as possible to previously hard coded export. But I did allow some deviation to creep in where it made more sense or was simply easier to implement in template language:

- Individual sprite frames are always exported into separate lines. Originally they were added to the same line until 10 frames were crossed and only then in new line.
- Collisions comment is not exported for tilemap as there is no data exported either. Originally it was always exported.

Note: in order to compensate for above changes, I had to tweak couple unit test template files.