Preparing for different types of source files.
This commit introduces a `SourceFile` base class and makes `SourceImage` a subclass of it.

Note: unfortunately there is one instance where we need to call specific `SourceImage` method directly from outside, so I had to implement it using type testing pattern matching. Not the best, but good enough for now...