Model's `BlocksAcross` is now automatically updated when `GridWidth` changes.
This follows the "don't ask, we'll ask you" principle (or in other words "push not pull" :) Most often than not, that's better than requiring external objects to know when certain data needs to change. Though in this case we still have to rely on external object to provide us bitmap width. Maybe it would be a good idea to handle the bitmap internally?