Changed settings form to non-modal.
This allows on-the-fly changes. However since user can now click remap or export at any time while settings form is displayed, any change here must be immediately reflected in the model. This requires adding event handlers to each and every control on settings form. To make it simpler, reflection is used which allows one line setup. Simple to use but may be a bit daunting.