Improved main form tiles/blocks images handling.
The images are now contained within split container so user can increase/decrease each one separately. It's also possible to increase/decrease the size of the right panel through splitter. However it's no longer possible to quickly scroll vertically (as it was before). But I think the combination of splitters and scrollable `ImageForm` should compensate.