Implemented tilemap preview.
This also adds ability for `IndexedBitmap` to render itself flipped and/or rotated. 

Note: `IndexedBitmap.CopyTo` method was changed to use `Point` instead of `Rectangle` when rendering. Rectangle usage was confusing since the same object was containing A) location to render into on destination side and B) size to render from source (as well as destination). This required caller to ensure size was correct. But we always rendered whole objects, and because the bitmap already knows its size, `Point` makes clearer API.