Palette form is shown if needed without prior dialog.
This should be better UX - user immediately has the palette shown so they can make informative decision instead of yes/no dialog without any visual indication of what palette actually looks like. The dialog was not really needed, it just resulted in extra click (if user chose to open it). On the hand, it's still as easy to dismiss with single key press, as it was with the dialog if user chooses not to change palette.